Hoist ConnectionPool out of SchemaLoader up to PostgresConnectorSup
