+++ components/electric/lib/electric/postgres/extension/schema_loader/epgsql.ex
@@ -6,106 +6,37 @@ defmodule Electric.Postgres.Extension.SchemaLoader.Epgsql do
   Uses a connection pool to avoid deadlocks when e.g. refreshing a subscription
   then attempting to run a query against the db.
   """
-  defmodule ConnectionPool do
-    @moduledoc false
-
-    alias Electric.Replication.Connectors
-
-    require Logger
-
-    @behaviour NimblePool
-
-    @impl NimblePool
-    def init_worker(conn_config) do
-      Logger.debug("Starting SchemaLoader pg connection: #{inspect(conn_config)}")
-      # NOTE: use `__connection__: conn` in tests to pass an existing connection
-      {:ok, conn} =
-        case Keyword.fetch(conn_config, :__connection__) do
-          {:ok, conn} ->
-            {:ok, conn}
-
-          :error ->
-            conn_config
-            |> Connectors.get_connection_opts(replication: false)
-            |> :epgsql.connect()
-        end
-
-      {:ok, conn, conn_config}
-    end
-
-    @impl NimblePool
-    # Transfer the port to the caller
-    def handle_checkout(:checkout, _from, conn, pool_state) do
-      {:ok, conn, conn, pool_state}
-    end
-
-    @impl NimblePool
-    def handle_checkin(:ok, _from, conn, pool_state) do
-      {:ok, conn, pool_state}
-    end
-
-    @impl NimblePool
-    def terminate_worker(_reason, conn, pool_state) do
-      Logger.debug("Terminating idle db connection #{inspect(conn)}")
-      :epgsql.close(conn)
-      {:ok, pool_state}
-    end
-
-    @impl NimblePool
-    def handle_ping(_conn, _pool_state) do
-      {:remove, :idle}
-    end
-  end
 
   alias Electric.Postgres.{Extension, Extension.SchemaLoader, Schema}
   alias Electric.Replication.{Connectors, Postgres.Client}
+  alias Electric.Postgres.ConnectionPool
 
   require Logger
 
   @behaviour SchemaLoader
 
-  @pool_timeout 5_000
-
   @impl true
   def connect(conn_config, _opts) do
-    {:ok, _pool} =
-      NimblePool.start_link(
-        worker: {ConnectionPool, conn_config},
-        # only connect when required, not immediately
-        lazy: true,
-        pool_size: 4,
-        worker_idle_timeout: 30_000
-      )
-  end
-
-  defp checkout!(pool, fun) do
-    NimblePool.checkout!(
-      pool,
-      :checkout,
-      fn _pool, conn ->
-        {fun.(conn), :ok}
-      end,
-      @pool_timeout
-    )
+    {:ok, ConnectionPool.name(Connectors.origin(conn_config))}
   end
 
   @impl true
   def load(pool) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       Extension.current_schema(conn)
     end)
   end
 
   @impl true
   def load(pool, version) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       Extension.schema_version(conn, version)
     end)
   end
 
   @impl true
   def save(pool, version, schema, stmts) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       with :ok <- Extension.save_schema(conn, version, schema, stmts) do
         {:ok, pool}
       end
@@ -118,7 +49,7 @@ defmodule Electric.Postgres.Extension.SchemaLoader.Epgsql do
   end
 
   def relation_oid(pool, rel_type, schema, table) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       Client.relation_oid(conn, rel_type, schema, table)
     end)
   end
@@ -138,7 +69,7 @@ defmodule Electric.Postgres.Extension.SchemaLoader.Epgsql do
 
   @impl true
   def primary_keys(pool, schema, name) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       {:ok, _, pks_data} = :epgsql.equery(conn, @primary_keys_query, [schema, name])
 
       {:ok, Enum.map(pks_data, &elem(&1, 0))}
@@ -147,14 +78,12 @@ defmodule Electric.Postgres.Extension.SchemaLoader.Epgsql do
 
   @impl true
   def primary_keys(pool, {schema, name}) do
-    checkout!(pool, fn conn ->
-      primary_keys(conn, schema, name)
-    end)
+    primary_keys(pool, schema, name)
   end
 
   @impl true
   def refresh_subscription(pool, name) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       query = ~s|ALTER SUBSCRIPTION "#{name}" REFRESH PUBLICATION WITH (copy_data = false)|
 
       case :epgsql.squery(conn, query) do
@@ -175,21 +104,21 @@ defmodule Electric.Postgres.Extension.SchemaLoader.Epgsql do
 
   @impl true
   def migration_history(pool, version) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       Extension.migration_history(conn, version)
     end)
   end
 
   @impl true
   def known_migration_version?(pool, version) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       Extension.known_migration_version?(conn, version)
     end)
   end
 
   @impl true
   def internal_schema(pool) do
-    checkout!(pool, fn conn ->
+    ConnectionPool.checkout!(pool, fn conn ->
       oid_loader = &Client.relation_oid(conn, &1, &2, &3)
 
       Enum.reduce(Extension.replicated_table_ddls(), Schema.new(), fn ddl, schema ->
